rename: auth 구조 변경(#7)
- Login.jsx, Signup.jsx, Dashboard.jsx를 새로 생성한 'auth' 폴더로 이동
- App.jsx의 파일 import 주소 변경

해결: #7
